[IMP] *: bigint test

+++ odoo/addons/test_new_api/tests/test_properties.py
@@ -1312,7 +1312,7 @@ class PropertiesCase(TestPropertiesMixin):
             UPDATE "test_new_api_message"
             SET "attributes" = "__tmp"."attributes"::jsonb,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "attributes", "write_date", "write_uid")
             WHERE "test_new_api_message"."id" = "__tmp"."id"
         """]
